refactor: align about presentation package with its directory
Keep feature About and support Compose screens in
about.presentation. Shared scaffolds stay in ui.compose.
