+++ app/src/main/java/com/dpis/module/fonts/FontDetailActivity.kt
@@ -14,7 +14,7 @@ import com.dpis.module.ui.compose.FontDetailDialog
 import com.dpis.module.ui.compose.FontDetailPresentation
 import com.dpis.module.ui.compose.FontDetailUiState
 import com.dpis.module.ui.compose.FontReferenceUiItem
-import com.dpis.module.ui.compose.SupportActivityContent
+import com.dpis.module.about.presentation.SupportActivityContent
 
 class FontDetailActivity : LocalizedActivity() {
     private lateinit var fontLibraryStore: FontLibraryStore
