+++ app/src/main/java/com/dpis/module/fonts/FontLibraryActivity.kt
@@ -14,7 +14,7 @@ import com.dpis.module.runtime.RuntimeConfigDelivery
 import com.dpis.module.ui.compose.FontLibraryDialog
 import com.dpis.module.ui.compose.FontLibraryPresentation
 import com.dpis.module.ui.compose.FontLibraryUiItem
-import com.dpis.module.ui.compose.SupportActivityContent
+import com.dpis.module.about.presentation.SupportActivityContent
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
