+++ app/src/main/java/com/dpis/module/settings/ThemeSettingsActivity.kt
@@ -2,7 +2,7 @@ package com.dpis.module.settings
 
 import android.os.Bundle
 import com.dpis.module.settings.LocalizedActivity
-import com.dpis.module.ui.compose.SupportActivityContent
+import com.dpis.module.about.presentation.SupportActivityContent
 
 class ThemeSettingsActivity : LocalizedActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
